+++ mdw-services/src/com/centurylink/mdw/listener/jms/ExternalEventListener.java
@@ -15,11 +15,6 @@
  */
 package com.centurylink.mdw.listener.jms;
 
-/*
- * Copyright (c) 2011 CenturyLink, Inc. All Rights Reserved.
- */
-
-
 import java.util.HashMap;
 import java.util.Map;
 
@@ -34,14 +29,9 @@ import javax.jms.Session;
 import javax.jms.TextMessage;
 
 import com.centurylink.mdw.container.ThreadPoolProvider;
-import com.centurylink.mdw.dataaccess.DataAccessException;
 import com.centurylink.mdw.listener.ListenerHelper;
 import com.centurylink.mdw.model.listener.Listener;
-import com.centurylink.mdw.model.monitor.CertifiedMessage;
-import com.centurylink.mdw.services.EventManager;
-import com.centurylink.mdw.services.ServiceLocator;
 import com.centurylink.mdw.util.JMSServices;
-import com.centurylink.mdw.util.ServiceLocatorException;
 import com.centurylink.mdw.util.log.LoggerUtil;
 import com.centurylink.mdw.util.log.StandardLogger;
 
@@ -51,45 +41,8 @@ public class ExternalEventListener extends JmsListener {
         super(ThreadPoolProvider.WORKER_LISTENER, queueName, thread_pool);
     }
 
-    @Override
-    protected TextMessage filterMessage(TextMessage message) throws DataAccessException, ServiceLocatorException, JMSException {
-        String msgid = message.getJMSCorrelationID();
-        if (isCertifiedMessage(msgid)) {
-            EventManager eventManager = ServiceLocator.getEventManager();
-            boolean consumed = eventManager.consumeCertifiedMessage(msgid);
-            acknowledge(message, msgid);
-            return consumed?message:null;
-        }
-        return message;
-    }
-
-    private void acknowledge(Message jmsMessage, String msgid) throws JMSException, ServiceLocatorException {
-        QueueConnection connection = null;
-        QueueSession session = null;
-        QueueSender sender = null;
-        try {
-            Queue respQueue = (Queue) jmsMessage.getJMSReplyTo();
-            QueueConnectionFactory qcf = JMSServices.getInstance().getQueueConnectionFactory(null);
-            connection = qcf.createQueueConnection();
-            session = connection.createQueueSession(false, QueueSession.DUPS_OK_ACKNOWLEDGE);
-            sender = session.createSender(respQueue);
-            Message respMsg = session.createTextMessage(msgid);
-//        respMsg.setJMSCorrelationID(correlationId); not used
-            sender.send(respMsg);
-        } finally {
-            if (sender != null) sender.close();
-            if (session != null) session.close();
-            if (connection != null) connection.close();
-        }
-    }
-
     protected Runnable getProcesser(TextMessage message) throws JMSException {
         return new ExternalEventDriver(message);
-    }
-
-    private boolean isCertifiedMessage(String correlationId) {
-        return (correlationId!=null &&
-                correlationId.startsWith(CertifiedMessage.CERTIFIED_MESSAGE_PREFIX));
     }
 
     class ExternalEventDriver implements Runnable {
@@ -118,7 +71,7 @@ public class ExternalEventListener extends JmsListener {
                     resp = helper.processEvent(txt, metaInfo);
                     Queue respQueue = (Queue) message.getJMSReplyTo();
                     String correlId = message.getJMSCorrelationID();
-                    if (resp != null && respQueue != null && !isCertifiedMessage(correlId)) {
+                    if (resp != null && respQueue != null) {
                         // String msgId = jmsMessage.getJMSMessageID();
                         QueueConnectionFactory qcf
                             = JMSServices.getInstance().getQueueConnectionFactory(null);
