+++ mdw-services/src/com/centurylink/mdw/services/event/EventManagerBean.java
@@ -22,35 +22,24 @@ import java.util.Date;
 import java.util.List;
 import java.util.Map;
 
-import org.apache.xmlbeans.XmlException;
-import org.apache.xmlbeans.XmlObject;
-
 import com.centurylink.mdw.app.ApplicationContext;
-import com.centurylink.mdw.cache.impl.PackageCache;
 import com.centurylink.mdw.common.MdwException;
-import com.centurylink.mdw.connector.adapter.AdapterException;
-import com.centurylink.mdw.connector.adapter.ConnectionException;
 import com.centurylink.mdw.constant.ActivityResultCodeConstant;
 import com.centurylink.mdw.container.ThreadPoolProvider;
 import com.centurylink.mdw.dataaccess.DataAccess;
 import com.centurylink.mdw.dataaccess.DataAccessException;
 import com.centurylink.mdw.dataaccess.DatabaseAccess;
 import com.centurylink.mdw.dataaccess.RuntimeDataAccess;
-import com.centurylink.mdw.event.EventHandler;
 import com.centurylink.mdw.model.Response;
 import com.centurylink.mdw.model.event.EventInstance;
 import com.centurylink.mdw.model.event.EventLog;
 import com.centurylink.mdw.model.event.EventType;
 import com.centurylink.mdw.model.event.InternalEvent;
-import com.centurylink.mdw.model.listener.Listener;
-import com.centurylink.mdw.model.monitor.CertifiedMessage;
 import com.centurylink.mdw.model.monitor.ScheduledEvent;
 import com.centurylink.mdw.model.monitor.UnscheduledEvent;
-import com.centurylink.mdw.model.request.Request;
 import com.centurylink.mdw.model.user.UserAction;
 import com.centurylink.mdw.model.user.UserAction.Action;
 import com.centurylink.mdw.model.variable.Document;
-import com.centurylink.mdw.model.variable.DocumentReference;
 import com.centurylink.mdw.model.variable.Variable;
 import com.centurylink.mdw.model.variable.VariableInstance;
 import com.centurylink.mdw.model.workflow.Activity;
@@ -67,14 +56,9 @@ import com.centurylink.mdw.services.EventException;
 import com.centurylink.mdw.services.EventManager;
 import com.centurylink.mdw.services.ProcessException;
 import com.centurylink.mdw.services.messenger.InternalMessenger;
-import com.centurylink.mdw.services.messenger.IntraMDWMessenger;
 import com.centurylink.mdw.services.messenger.MessengerFactory;
-import com.centurylink.mdw.services.pooling.AdapterConnectionPool;
-import com.centurylink.mdw.services.pooling.ConnectionPoolRegistration;
-import com.centurylink.mdw.services.pooling.PooledAdapterConnection;
 import com.centurylink.mdw.services.process.InternalEventDriver;
 import com.centurylink.mdw.services.process.ProcessExecutor;
-import com.centurylink.mdw.translator.VariableTranslator;
 import com.centurylink.mdw.util.TransactionWrapper;
 import com.centurylink.mdw.util.log.LoggerUtil;
 import com.centurylink.mdw.util.log.StandardLogger;
@@ -147,93 +131,6 @@ public class EventManagerBean implements EventManager {
             return edao.getDistinctEventLogEventSources();
         } catch (SQLException e) {
             throw new EventException("Failed to notify events", e);
-        } finally {
-            edao.stopTransaction(transaction);
-        }
-    }
-
-    public String processExternalEvent(String clsname, String request, Map<String,String> metainfo)
-        throws Exception {
-        String packageName = metainfo.get(Listener.METAINFO_PACKAGE_NAME);
-        Package pkg = PackageCache.getPackage(packageName);
-        EventHandler handler = pkg.getEventHandler(clsname, request, metainfo);
-        XmlObject xmlBean = XmlObject.Factory.parse(request);
-        Request requestDoc = new Request(0L);
-        requestDoc.setContent(request);
-        return handler.handleEventMessage(requestDoc, xmlBean, metainfo).getContent();
-    }
-
-    /**
-     * Creates the process instance. This is only used by TaskManger to create
-     * a dummy process instance for tasks not associated with process instances.
-     *
-     * @param pProcessId
-     * @param pProcessOwner
-     * @param pProcessOwnerId
-     * @param pSecondaryOwner
-     * @param pSecondaryOwnerId
-     * @return new WorkInstance
-     */
-    public ProcessInstance createProcessInstance(Long pProcessId, String pProcessOwner,
-        Long pProcessOwnerId, String pSecondaryOwner, Long pSecondaryOwnerId,
-        String pMasterRequestId)
-    throws ProcessException, DataAccessException {
-        TransactionWrapper transaction = null;
-        EngineDataAccessDB edao = new EngineDataAccessDB();
-        try {
-            transaction = edao.startTransaction();
-            Process processVO = ProcessCache.getProcess(pProcessId);
-            ProcessInstance pi = new ProcessInstance(pProcessId, processVO.getName());
-            pi.setOwner(pProcessOwner);
-            pi.setOwnerId(pProcessOwnerId);
-            pi.setSecondaryOwner(pSecondaryOwner);
-            pi.setSecondaryOwnerId(pSecondaryOwnerId);
-            pi.setMasterRequestId(pMasterRequestId);
-            pi.setStatusCode(WorkStatus.STATUS_PENDING_PROCESS);
-            edao.createProcessInstance(pi);
-            return pi;
-        } catch (Exception e) {
-            throw new ProcessException(-1, e.getMessage(), e);
-        } finally {
-            edao.stopTransaction(transaction);
-        }
-
-    }
-
-    /**
-     * updates the variable instance. If the variable type is a document type,
-     * this method updates the content of the document.
-     *
-     * @param pVarInstId
-     * @param pVariableData data to be updated. If the variable is a document variable,
-     *         this should be the actual content
-     * @throws com.centurylink.mdw.dataaccess.DataAccessException
-     */
-    public void updateVariableInstance(Long pVarInstanceId, Object pVariableData)
-    throws ProcessException, com.centurylink.mdw.dataaccess.DataAccessException {
-        TransactionWrapper transaction = null;
-        EngineDataAccessDB edao = new EngineDataAccessDB();
-        try {
-            transaction = edao.startTransaction();
-            VariableInstance varInst = edao.getVariableInstance(pVarInstanceId);
-            // do we need to check if this also looks for variables defined in ancestor processes?
-            if (VariableTranslator.isDocumentReferenceVariable(varInst.getType())) {
-                DocumentReference docref = (DocumentReference)varInst.getData();
-                Document docvo = edao.getDocument(docref.getDocumentId(), false);
-                if (pVariableData instanceof String)
-                    docvo.setContent((String)pVariableData);
-                else
-                    docvo.setObject(pVariableData, varInst.getType());
-                edao.updateDocumentContent(docvo.getDocumentId(), docvo.getContent());
-            } else {
-                if (pVariableData instanceof String)
-                    varInst.setStringValue((String)pVariableData);
-                else
-                    varInst.setData(pVariableData);
-                edao.updateVariableInstance(varInst);
-            }
-        } catch (SQLException e) {
-            throw new DataAccessException(-1, "Failed to update document content", e);
         } finally {
             edao.stopTransaction(transaction);
         }
@@ -373,32 +270,6 @@ public class EventManagerBean implements EventManager {
     }
 
 
-    /**
-     * Sends the JMS Message to the EventHandler
-     *
-     * @param pMsg Message to be send out
-     * @param pDelay delay the delivery of the message in seconds
-     * @throws ProcessException
-     */
-    public void sendInternalEvent(String message) throws ProcessException {
-        TransactionWrapper transaction = null;
-        EngineDataAccessDB edao = new EngineDataAccessDB();
-        try {
-            transaction = edao.startTransaction();
-            this.sendInternalEvent(new InternalEvent(message), edao);
-        } catch (DataAccessException e) {
-            throw new ProcessException(0, "Failed to send internal message", e);
-        } catch (XmlException e) {
-            throw new ProcessException(0, "Failed to send internal message", e);
-        } finally {
-            try {
-                edao.stopTransaction(transaction);
-            } catch (DataAccessException e) {
-                throw new ProcessException(0, "Failed to send internal message", e);
-            }
-        }
-    }
-
     private void sendInternalEvent(InternalEvent pMsg, EngineDataAccess edao) throws ProcessException {
         InternalMessenger msgbroker = MessengerFactory.newInternalMessenger();
         msgbroker.sendMessage(pMsg, edao);
@@ -525,24 +396,6 @@ public class EventManagerBean implements EventManager {
     }
 
     /**
-     * Returns the ProcessInstance and associated variable  identified by the passed in Id
-     *
-     * @param pProcInstId
-     * @param loadVariables
-     * @return ProcessInstance
-     */
-    public ProcessInstance getProcessInstance(Long procInstId, boolean loadVariables)
-    throws ProcessException, DataAccessException {
-        if (loadVariables)
-        {
-            ProcessInstance processInstanceVO = this.getProcessInstance(procInstId);
-            List<VariableInstance> variables =  this.getProcessInstanceVariables(procInstId);
-            processInstanceVO.setVariables(variables);
-            return processInstanceVO;
-        }
-        return this.getProcessInstance(procInstId);
-    }
-    /**
      * Returns the process instances by process name and master request ID.
      *
      * @param processName
@@ -655,11 +508,6 @@ public class EventManagerBean implements EventManager {
         return wti;
     }
 
-    public void updateDocumentContent(Long docid, Object doc, String type)
-    throws DataAccessException {
-        updateDocumentContent(docid, doc, type, null);
-    }
-
     public void updateDocumentContent(Long docid, Object doc, String type, Package pkg)
     throws DataAccessException {
         TransactionWrapper transaction = null;
@@ -701,11 +549,6 @@ public class EventManagerBean implements EventManager {
         finally {
             edao.stopTransaction(transaction);
         }
-    }
-
-    public Long createDocument(String type, String ownerType, Long ownerId, Object doc)
-    throws DataAccessException {
-        return createDocument(type, ownerType, ownerId, doc, null);
     }
 
     public Long createDocument(String type, String ownerType, Long ownerId, Object doc, Package pkg)
@@ -903,274 +746,6 @@ public class EventManagerBean implements EventManager {
     }
 
 
-    public List<CertifiedMessage> getCertifiedMessageList()
-    throws DataAccessException {
-        TransactionWrapper transaction = null;
-        EngineDataAccessDB edao = new EngineDataAccessDB();
-        try {
-            transaction = edao.startTransaction();
-            return edao.getCertifiedMessageList();
-        } catch (SQLException e) {
-            throw new DataAccessException(-1, "Failed to update document content", e);
-        } finally {
-            edao.stopTransaction(transaction);
-        }
-    }
-
-    public void recordCertifiedMessage(CertifiedMessage message)
-    throws DataAccessException {
-        TransactionWrapper transaction = null;
-        EngineDataAccessDB edao = new EngineDataAccessDB();
-        try {
-            transaction = edao.startTransaction();
-            edao.recordCertifiedMessage(message);
-        } catch (SQLException e) {
-            throw new DataAccessException(-1, "Failed to process delayed event", e);
-        } finally {
-            edao.stopTransaction(transaction);
-        }
-    }
-
-    /**
-     *
-     * @param message
-     * @param ackTimeout timeout in seconds for waiting for acknowledgment
-     * @param maxTries maximum number of tries
-     * @param retryInterval elapse time in seconds for the next retry
-     * @return true when retry is needed, false o/w
-     * @throws DataAccessException
-     */
-    public boolean deliverCertifiedMessage(CertifiedMessage message, int ackTimeout, int maxTries,
-            int retryInterval) {
-        TransactionWrapper transaction = null;
-        EngineDataAccessDB edao = new EngineDataAccessDB();
-        try {
-            transaction = edao.startTransaction();
-            // 1. lock the certified message
-            CertifiedMessage refreshed;
-            try {
-                refreshed = edao.lockCertifiedMessage(message.getId());
-            } catch (SQLException e) {
-                logger.severe("Failed to lock certified message " + message.getId() + ", retry again");
-                message.setNextTryTime(new Date(DatabaseAccess.getCurrentTime()+1000*retryInterval));
-                return true;
-            }
-            // 2. handle the case when the message has been deleted (only possible to delete from admin
-            if (refreshed==null) {
-                // only possible when the message is deleted from admin
-                message.setStatus(EventInstance.STATUS_CERTIFIED_MESSAGE_CANCEL);
-                return false;
-            }
-            // 3. handle the case when the message has been processed by another managed server
-            int count = refreshed.getTryCount();
-            if (!refreshed.getStatus().equals(EventInstance.STATUS_CERTIFIED_MESSAGE)) return false;
-            if (count!=message.getTryCount()) {
-                // the message has been retried by some other server, successful or failed
-                message.setTryCount(count);
-                if (refreshed.getNextTryTime()!=null) {
-                    message.setNextTryTime(refreshed.getNextTryTime());
-                    return true;
-                } else return false;
-            }
-            // 4. load message content if not already in memory
-            if (message.getContent()==null) {
-                try {
-                    Document docvo = this.getDocumentVO(message.getDocumentId());
-                    message.setContent(docvo.getContent());
-                } catch (DataAccessException e) {
-                    message.setStatus(EventInstance.STATUS_CERTIFIED_MESSAGE_HOLD);
-                    logger.severeException(LoggerUtil.getStandardLogger().getSentryMark()+
-                            "Failed to load certified message content: " + message.getId(), e);
-                    updateCertifiedMessageStatus(edao, message.getId(), message.getStatus(), 0,
-                            new Date(DatabaseAccess.getCurrentTime()));
-                    return false;
-                }
-            }
-            int exceptionCode = 0;
-            String protocol = message.getProperty(CertifiedMessage.PROP_PROTOCOL);
-            PooledAdapterConnection conn = null;
-            try {
-                if (CertifiedMessage.PROTOCOL_POOL.equals(protocol)) {
-                    // 5a. handle connection pool based certified messages
-                    String pool_name = message.getProperty(CertifiedMessage.PROP_POOL_NAME);
-                    if (pool_name==null) throw new AdapterException("Pool name is not specified");
-                    AdapterConnectionPool pool = ConnectionPoolRegistration.getPool(pool_name);
-                    if (pool==null) throw new AdapterException("Connection pool is not configured: " + pool_name);
-                    conn = pool.getConnection("Certified Message Manager", null);
-                    conn.invoke(message.getContent(), ackTimeout, message.getProperties());
-                } else {
-                    // 5b. handle JMS based certified messages
-                    String jndi = message.getProperty(CertifiedMessage.PROP_JNDI_URL);
-                    IntraMDWMessenger messenger = MessengerFactory.newIntraMDWMessenger(jndi);
-                    messenger.sendCertifiedMessage(message.getContent(), message.getId(), ackTimeout);
-                }
-                // 6a. mark message delivered
-                message.setStatus(EventInstance.STATUS_CERTIFIED_MESSAGE_DELIVERED);
-                updateCertifiedMessageStatus(edao, message.getId(), message.getStatus(), 0,
-                        new Date(DatabaseAccess.getCurrentTime()));
-                return false;
-            } catch (AdapterException e1) {
-                // 6b. handle non-retriable errors
-                exceptionCode = ((AdapterException)e1).getCode();
-                message.setStatus(EventInstance.STATUS_CERTIFIED_MESSAGE_HOLD);
-                logger.severeException(LoggerUtil.getStandardLogger().getSentryMark()+ "Certified message hits unretriable error: "
-                            + message.getId(), e1);
-                updateCertifiedMessageStatus(edao, message.getId(), message.getStatus(), 0,
-                        new Date(DatabaseAccess.getCurrentTime()));
-                return false;
-            } catch (Exception e1) {
-                // 6c. handle retriable errors
-                // this includes the case for NoSuchElementException - pool connections are exhausted
-                String exceptionName = e1.getClass().getName();
-                if (e1 instanceof ConnectionException) {
-                    exceptionCode = ((ConnectionException)e1).getCode();
-                    exceptionName += ":" + exceptionCode;
-                }
-                if (count>0) {
-                    logger.severe("Failed to deliver certified message " + message.getId()
-                            + ", exception " + exceptionName
-                            + " - retry " + count);
-                } else {
-                    logger.severeException("Failed to deliver certified message " + message.getId(), e1);
-                }
-                if (exceptionCode==ConnectionException.POOL_DISABLED) {
-                    logger.severe("Pool is disabled - indefinite wait for certified message " + message.getId());
-                    updateCertifiedMessageStatus(edao, message.getId(), message.getStatus(), 0, null);
-                    return false;
-                } else if (count>=maxTries) {
-                    // 7a. when exceeding max retry limit
-                    message.setStatus(EventInstance.STATUS_CERTIFIED_MESSAGE_HOLD);
-                    logger.severe(LoggerUtil.getStandardLogger().getSentryMark() +
-                            "Certified message failed to deliver after max tries: " + message.getId());
-                    updateCertifiedMessageStatus(edao, message.getId(), message.getStatus(), 0,
-                            new Date(DatabaseAccess.getCurrentTime()));
-                    return false;
-                } else {
-                    // 7b. when retry count is within limit, really retrying
-                    message.setTryCount(count+1);
-                    message.setNextTryTime(new Date(DatabaseAccess.getCurrentTime()+1000*retryInterval));
-                    updateCertifiedMessageStatus(edao, message.getId(),
-                            EventInstance.STATUS_CERTIFIED_MESSAGE, message.getTryCount(), message.getNextTryTime());
-                    return true;
-                }
-            } finally {
-                if (exceptionCode==0) exceptionCode = -1;
-                if (conn!=null) conn.returnConnection(exceptionCode);
-            }
-        } catch (DataAccessException e) {
-            logger.severe("Failed to start transaction for delivering certified message " + message.getId() + ", retry again");
-            message.setNextTryTime(new Date(DatabaseAccess.getCurrentTime()+1000*retryInterval));
-            return true;
-        } finally {
-            try {
-                edao.stopTransaction(transaction);
-            } catch (DataAccessException e) {
-                logger.severeException("Fail to commit transaction", e);
-            }
-        }
-    }
-
-    private void updateCertifiedMessageStatus(EngineDataAccessDB edao, String msgid,
-            Integer status, int tryCount, Date consumeTime) {
-        try {
-            edao.updateCertifiedMessageStatus(msgid, status, tryCount, consumeTime);
-        } catch (SQLException e) {
-            logger.severeException("Failed to update certified message status in database", e);
-        }
-    }
-
-    public boolean consumeCertifiedMessage(String messageId)
-    throws DataAccessException {
-        TransactionWrapper transaction = null;
-        EngineDataAccessDB edao = new EngineDataAccessDB();
-        try {
-            transaction = edao.startTransaction();
-            edao.consumeCertifiedMessage(messageId);
-            return true;
-        } catch (SQLIntegrityConstraintViolationException e) {
-            return false;
-        } catch (SQLException e) {
-            if (e.getSQLState().equals("23000")) return false;
-            throw new DataAccessException(-1, "Failed to consume certified message", e);
-        } finally {
-            edao.stopTransaction(transaction);
-        }
-    }
-
-    public void updateCertifiedMessageStatus(String msgid, Integer status)
-    throws DataAccessException {
-        TransactionWrapper transaction = null;
-        EngineDataAccessDB edao = new EngineDataAccessDB();
-        try {
-            transaction = edao.startTransaction();
-            edao.updateCertifiedMessageStatus(msgid, status, 0,
-                    new Date(DatabaseAccess.getCurrentTime()));
-        } catch (Exception e) {
-            throw new DataAccessException(-1, e.getMessage(), e);
-        } finally {
-            edao.stopTransaction(transaction);
-        }
-    }
-
-    public int getTableRowCount(String tableName, String whereClause) throws DataAccessException {
-        TransactionWrapper transaction = null;
-        EngineDataAccessDB edao = new EngineDataAccessDB();
-        try {
-            transaction = edao.startTransaction();
-            return edao.getTableRowCount(tableName, whereClause);
-        } catch (SQLException e) {
-            throw new DataAccessException(-1, "Failed to obtain event instance count", e);
-        } finally {
-            edao.stopTransaction(transaction);
-        }
-    }
-
-    public List<String[]> getTableRowList(String tableName, Class<?>[] types, String[] fields,
-            String whereClause, String orderby, boolean descending, int startRow, int rowCount)
-    throws DataAccessException {
-        TransactionWrapper transaction = null;
-        EngineDataAccessDB edao = new EngineDataAccessDB();
-        try {
-            transaction = edao.startTransaction();
-            return edao.getTableRowList(tableName, types, fields, whereClause, orderby, descending, startRow, rowCount);
-        } catch (SQLException e) {
-            throw new DataAccessException(-1, "Failed to retrieve event instance list", e);
-        } finally {
-            edao.stopTransaction(transaction);
-        }
-    }
-
-    public int deleteTableRow(String tableName, String fieldName, Object fieldValue) throws DataAccessException {
-        TransactionWrapper transaction = null;
-        EngineDataAccessDB edao = new EngineDataAccessDB();
-        try {
-            transaction = edao.startTransaction();
-            return edao.deleteTableRow(tableName, fieldName, fieldValue);
-        } catch (SQLException e) {
-            throw new DataAccessException(-1, "Failed to delete " + fieldValue.toString() + " from " + tableName, e);
-        } finally {
-            edao.stopTransaction(transaction);
-        }
-    }
-
-    public void createEventInstance(String eventName,
-            Long documentId, Integer status, Date consumeDate, String auxdata, String reference, int preserveSeconds)
-    throws DataAccessException {
-        TransactionWrapper transaction = null;
-        EngineDataAccessDB edao = new EngineDataAccessDB();
-        try {
-            transaction = edao.startTransaction();
-            edao.createEventInstance(eventName, documentId, status, consumeDate, auxdata, reference, preserveSeconds);
-        } catch (SQLIntegrityConstraintViolationException e) {
-            throw new DataAccessException(23000, "The event instance already exists", e);
-        } catch (SQLException e) {
-            if (e.getSQLState().equals("23000"))
-                throw new DataAccessException(23000, "The event is already scheduled", e);
-            else throw new DataAccessException(-1, "Failed to retrieve event instance list", e);
-        } finally {
-            edao.stopTransaction(transaction);
-        }
-    }
 
     public void updateEventInstance(String eventName,
             Long documentId, Integer status, Date consumeDate, String auxdata, String reference, int preserveSeconds)
@@ -1188,49 +763,6 @@ public class EventManagerBean implements EventManager {
             edao.updateEventInstance(eventName, documentId, status, consumeDate, auxdata, reference, preserveSeconds, comments);
         } catch (SQLException e) {
             throw new DataAccessException(-1, "Failed to retrieve event instance list", e);
-        } finally {
-            edao.stopTransaction(transaction);
-        }
-    }
-
-    public void createEventWaitInstance(String eventName,
-            Long actInstId,  String compCode)
-    throws DataAccessException {
-        TransactionWrapper transaction = null;
-        EngineDataAccessDB edao = new EngineDataAccessDB();
-        try {
-            transaction = edao.startTransaction();
-            edao.createEventWaitInstance(actInstId, eventName, compCode);
-        } catch (SQLException e) {
-            throw new DataAccessException(-1, "Failed to create event wait instance", e);
-        } finally {
-            edao.stopTransaction(transaction);
-        }
-    }
-
-    public void createTableRow(String tableName, String[] fieldNames, Object[] fieldValues)
-        throws DataAccessException {
-        TransactionWrapper transaction = null;
-        EngineDataAccessDB edao = new EngineDataAccessDB();
-        try {
-            transaction = edao.startTransaction();
-            edao.createTableRow(tableName, fieldNames, fieldValues);
-        } catch (SQLException e) {
-            throw new DataAccessException(-1, "Failed to create entry in " + tableName, e);
-        } finally {
-            edao.stopTransaction(transaction);
-        }
-    }
-
-    public int updateTableRow(String tableName, String keyName, Object keyValue,
-            String[] fieldNames, Object[] fieldValues)  throws DataAccessException {
-        TransactionWrapper transaction = null;
-        EngineDataAccessDB edao = new EngineDataAccessDB();
-        try {
-            transaction = edao.startTransaction();
-            return edao.updateTableRow(tableName, keyName, keyValue, fieldNames, fieldValues);
-        } catch (SQLException e) {
-            throw new DataAccessException(-1, "Failed to update " + keyName + " of " + tableName, e);
         } finally {
             edao.stopTransaction(transaction);
         }
@@ -1321,25 +853,6 @@ public class EventManagerBean implements EventManager {
             }
         }
         return null;
-    }
-    /**
-     * Returns Document VO
-     * @param forUpdate
-     * @throws com.centurylink.mdw.dataaccess.DataAccessException
-     */
-    public Document getDocument(DocumentReference docRef, boolean forUpdate)
-    throws ProcessException, com.centurylink.mdw.dataaccess.DataAccessException {
-        TransactionWrapper transaction = null;
-        EngineDataAccessDB edao = new EngineDataAccessDB();
-        try {
-            transaction = edao.startTransaction();
-            Document docvo = edao.getDocument(docRef.getDocumentId(), forUpdate);
-            return docvo;
-        } catch (SQLException e) {
-            throw new DataAccessException(-1, "Failed to get document content", e);
-        } finally {
-            edao.stopTransaction(transaction);
-        }
     }
 
     @Override
